maintain file encoding when appending to $profile
In powershell, the `>>` operator always encodes the data as UTF16-LE, and makes no attempt to match the encoding of the file that it is appending data to. This results in an extra byte being inserted in between every byte in the desired string when the string is appended to the file.

Using `| Out-File -Append -Encoding ASCII` instead of `>>` does not insert the extra bytes when appending to UFT8 encoded files.